Add copyright and move EncodingUtil to internals
