+++ src/main/java/net/dv8tion/jda/internal/utils/EncodingUtil.java
@@ -1,19 +1,26 @@
-package net.dv8tion.jda.api.utils;
+/*
+ * Copyright 2015-2019 Austin Keener, Michael Ritter, Florian Spieß, and the JDA contributors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package net.dv8tion.jda.internal.utils;
 
 import java.io.UnsupportedEncodingException;
 import java.net.URLEncoder;
 import java.util.stream.Collectors;
 
 public class EncodingUtil {
-    /**
-     * URL-Encodes the given String to UTF-8 after
-     * form-data specifications (space {@literal ->} +)
-     *
-     * @param  chars
-     *         The characters to encode
-     *
-     * @return The encoded String
-     */
     public static String encodeUTF8(String chars)
     {
         try
